pyproject.toml
[project]
name = "teacher-ai-agent"
version = "0.1.0"
description = "AI teacher with structured storyboard pipeline"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.32.0",
    "openai-whisper>=20250625",
    "torch>=2.11.0",
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
    "chromadb>=1.5.8",
    "httpx>=0.28.1",
    "moviepy>=2.2.1",
    "pillow>=11.3.0,<12",
    "python-multipart>=0.0.26",
    "python-dotenv>=1.2.2",
    "elevenlabs>=2.44.0",
]
